refactor(mobile): ViewModel을 전용 view-models 디렉토리로 분리
- TodoItemViewModel을 4개 query options에서 중복 정의 → view-models/로 통합
- FriendUserViewModel 도입으로 displayName 로직 일원화
- FriendTodoItemViewModel을 TodoItemViewModel로 통합
