+++ apps/mobile/src/features/todo/presentations/components/TodoList/CategoryTodoList.tsx
@@ -8,8 +8,8 @@ import times from 'es-toolkit/compat/times';
 import { useCallback } from 'react';
 import { ActivityIndicator } from 'react-native';
 import { InView } from 'react-native-intersection-observer';
-import type { TodoItemViewModel } from '../../queries/get-todos-infinite-query-options';
 import { getTodosInfiniteQueryOptions } from '../../queries/get-todos-infinite-query-options';
+import type { TodoItemViewModel } from '../../view-models/todo-item.view-model';
 import { TodoItem } from './TodoItem';
 
 interface CategoryTodoListProps {
