+++ apps/mobile/src/features/todo/presentations/queries/get-friend-todos-query-options.ts
@@ -1,21 +1,9 @@
 import { useTodoService } from '@src/bootstrap/providers/di-provider';
 import { unwrap } from '@src/shared/errors/result';
-import { formatTime } from '@src/shared/utils/date';
 import { queryOptions } from '@tanstack/react-query';
 
-import type { TodoItem } from '../../models/todo.model';
 import { TODO_QUERY_KEYS } from '../constants/todo-query-keys.constant';
-
-export interface FriendTodoItemViewModel extends TodoItem {
-  formattedTime: string | null;
-  color: string;
-}
-
-const toViewModel = (todo: TodoItem): FriendTodoItemViewModel => ({
-  ...todo,
-  formattedTime: todo.scheduledTime ? formatTime(todo.scheduledTime) : null,
-  color: todo.category.color,
-});
+import { toTodoItemViewModel } from '../view-models/todo-item.view-model';
 
 export const getFriendTodosQueryOptions = (friendUserId: string, date: string) => {
   const todoService = useTodoService();
@@ -31,7 +19,7 @@ export const getFriendTodosQueryOptions = (friendUserId: string, date: string) =
       return unwrap(result);
     },
     select: (data) => ({
-      todos: data.todos.map(toViewModel),
+      todos: data.todos.map(toTodoItemViewModel),
     }),
     staleTime: 30_000,
   });
